requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
python-dotenv>=0.19.0
python-multipart>=0.0.5
email-validator>=1.1.3
psycopg2-binary>=2.9.1